# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Freelance Marketplace API
  description: A comprehensive API for managing a freelance marketplace platform
  version: 1.0.0
  contact:
    name: API Support
    email: support@freelancehub.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://freelancehub.com/api
    description: Production server
security:
  - BearerAuth: []
  - SessionAuth: []
paths:
...
  /reviews:
    get:
      tags:
        - Reviews
      summary: Get project reviews
      description: Retrieve reviews for a specific project
      parameters:
        - name: projectId
          in: query
          required: true
          description: Project ID
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Project ID is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Reviews
      summary: Create review
      description: Create a review for a completed order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - orderId
                - rating
              properties:
                orderId:
                  type: string
                  example: "order_123"
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 5
                comment:
                  type: string
                  example: "Excellent work! Highly recommended."
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: Validation error or order not completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /webhooks/stripe:
    post:
      tags:
        - Webhooks
      summary: Stripe webhook handler
      description: Handle Stripe webhook events for payment processing
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Stripe webhook event payload
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
        '400':
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Webhook handler failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "user_123"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [USER, ADMIN]
          example: "USER"
        bio:
          type: string
          example: "Experienced full-stack developer"
        skills:
          type: array
          items:
            type: string
          example: ["React", "Node.js", "TypeScript"]
        rating:
          type: number
          minimum: 0
          maximum: 5
          example: 4.8
        totalReviews:
          type: integer
          minimum: 0
          example: 15
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Project:
      type: object
      properties:
        id:
          type: string
          example: "proj_123"
        title:
          type: string
          example: "E-commerce Website Development"
        description:
          type: string
          example: "Build a complete e-commerce website with modern technologies"
        price:
          type: number
          minimum: 0
          example: 2500
        category:
          type: string
          example: "Web Development"
        tags:
          type: array
          items:
            type: string
          example: ["React", "Node.js", "PostgreSQL"]
        images:
          type: array
          items:
            type: string
          example: ["https://example.com/image1.jpg"]
        status:
          type: string
          enum: [ACTIVE, INACTIVE, COMPLETED]
          example: "ACTIVE"
        featured:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        author:
          $ref: '#/components/schemas/User'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        _count:
          type: object
          properties:
            orders:
              type: integer
            reviews:
              type: integer
    Order:
      type: object
      properties:
        id:
          type: string
          example: "order_123"
        status:
          type: string
          enum: [PENDING, PAID, COMPLETED, CANCELLED]
          example: "PAID"
        totalAmount:
          type: number
          minimum: 0
          example: 2500
        paymentId:
          type: string
          example: "pi_123"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        buyer:
          $ref: '#/components/schemas/User'
        project:
          $ref: '#/components/schemas/Project'
        review:
          $ref: '#/components/schemas/Review'
    Review:
      type: object
      properties:
        id:
          type: string
          example: "review_123"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: "Excellent work! Highly recommended."
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        reviewer:
          $ref: '#/components/schemas/User'
        project:
          $ref: '#/components/schemas/Project'
        order:
          $ref: '#/components/schemas/Order'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          example: 10
        total:
          type: integer
          minimum: 0
          example: 100
        pages:
          type: integer
          minimum: 0
          example: 10
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Validation error"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token